Mapping and End-Game Form updates
The map now has 2 states; one that perpetuates the display of data across all user sessions and one that displays the data of just the current user session. Using session names that are strings functions best for map usage. The two maps can be selected on the home screen. At the end of each game the Dart Likert form is what opens and the built-in forms in the html files have been removed. Once form is submitted, the session map opens up. If player leaves game early and still fills out form, the session map will still open.